Hoàn Thành Brand và Category

+++ QLBanHang_API/Controllers/CategoryController.cs
@@ -24,6 +24,29 @@ namespace QLBanHang_API.Controllers
 			return Ok(categories);
 		}
 
+		[HttpGet("GetAllDetailCategory")]
+		public async Task<IActionResult> GetAllDetailCategory()
+		{
+			var categories = await _categoryService.GetAllDetailCategory();
+			return Ok(categories);
+		}
+
+		[HttpGet("GetFilteredCategories")]
+		public async Task<IActionResult> GetFilteredCategories([FromQuery] string searchQuery = "", [FromQuery] int page = 1, [FromQuery] int pageSize = 8, [FromQuery] string sortCriteria = "name", [FromQuery] bool isDescending = false)
+		{
+			var (categories, totalRecords) = await _categoryService.GetFilteredCategoriesAsync(page, pageSize, searchQuery, sortCriteria, isDescending);
+
+			return Ok(categories);
+		}
+
+		[HttpGet("TotalPagesCategory")]
+		public async Task<IActionResult> GetTotalPagesCategory([FromQuery] string searchQuery = "")
+		{
+			var totalRecords = await _categoryService.GetTotalCategoriesAsync(searchQuery);
+			var totalPages = (int)Math.Ceiling((double)totalRecords / 8); // Điều chỉnh số item trên mỗi trang nếu cần
+			return Ok(totalPages);
+		}
+
 		[HttpGet("{id}")]
 		public async Task<IActionResult> GetCategory(Guid id)
 		{
@@ -31,36 +54,39 @@ namespace QLBanHang_API.Controllers
 			return Ok(category);
 		}
 
-        [HttpPost("Create")]
-        [ValidateModel]
-        public async Task<IActionResult> CreateCategory([FromBody] CategoryDto category)
-        {
-            var success = await _categoryService.CreateCategoryAsync(category);
-            if (!success)
-            {
-                return BadRequest("Unable to create category."); 
-            }
-            return CreatedAtAction(nameof(GetCategory), new { id = category.CategoryId }, category);
-        }
-
-        [HttpPut("{id}")]
+		[HttpPost("Create")]
 		[ValidateModel]
-		public async Task<IActionResult> UpdateCategory(Guid id, [FromBody] CategoryDto category)
+		public async Task<IActionResult> CreateCategory([FromBody] CategoryDto category)
 		{
-			var existingCategory = await _categoryService.GetCategoryByIdAsync(id);
-			if (existingCategory == null)
+			if (await _categoryService.IsCategoryNameExistsAsync(category.CategoryName!))
 			{
-				return NotFound();
+				return BadRequest("Tên danh mục đã tồn tại.");
 			}
-			category.CategoryId = id; // Ensure the correct ID is used
-			var success = await _categoryService.UpdateCategoryAsync(category);
-            if (!success)
-            {
-                return BadRequest("Unable to update category."); // Trả về thông báo nếu không thành công
-            }
 
-            return Ok("Thanh cong");
-        }
+			var success = await _categoryService.CreateCategoryAsync(category);
+			if (!success)
+			{
+				return BadRequest("Unable to create category.");
+			}
+			return CreatedAtAction(nameof(GetCategory), new { id = category.CategoryId }, category);
+		}
+
+		[HttpPut]
+		[ValidateModel]
+		public async Task<IActionResult> UpdateCategory([FromBody] CategoryDto category)
+		{
+			if (await _categoryService.IsCategoryNameExistsAsync(category.CategoryName!))
+			{
+				return BadRequest("Tên danh mục đã tồn tại.");
+			}
+			var success = await _categoryService.UpdateCategoryAsync(category);
+			if (!success)
+			{
+				return BadRequest("Unable to update category."); // Trả về thông báo nếu không thành công
+			}
+
+			return Ok("Thanh cong");
+		}
 
 		[HttpDelete("{id}")]
 		public async Task<IActionResult> DeleteCategory(Guid id)
@@ -70,14 +96,19 @@ namespace QLBanHang_API.Controllers
 			{
 				return NotFound();
 			}
+			var productsUsingCategory = await _categoryService.GetProductsByCategoryIdAsync(id);
+			if (productsUsingCategory.Any())
+			{
+				return BadRequest("Danh mục này đang được sử dụng bởi các sản phẩm, không thể xóa!");
+			}
 
-            var success = await _categoryService.DeleteCategoryAsync(id);
-            if (!success)
-            {
-                return BadRequest("Unable to delete category."); // Trả về thông báo nếu không thành công
-            }
+			var success = await _categoryService.DeleteCategoryAsync(id);
+			if (!success)
+			{
+				return BadRequest("Unable to delete category."); // Trả về thông báo nếu không thành công
+			}
 
-            return Ok("Thanh cong");
-        }
+			return Ok("Thanh cong");
+		}
 	}
 }
